wip: Refactor reconciler function
