--- controllers/quay/features_test.go
@@ -1,1 +1,0 @@
-package controllers
